refractor: Move Volume Viewer package-to-be into it's own folder structure
