+++ src/App.jsx
@@ -3,10 +3,10 @@ import "aframe";
 import "aframe-event-set-component";
 import "aframe-orbit-controls";
 
-import "./Aframe/loader.js";
-import "./Aframe/buttons-check.js";
-import "./Aframe/render-2d-clipplane";
-import "./Aframe/cursor-listener";
+import "./VolumeViewer-package/Aframe/loader.js";
+import "./VolumeViewer-package/Aframe/buttons-check.js";
+import "./VolumeViewer-package/Aframe/cursor-listener";
+import "./VolumeViewer-package/Aframe/render-2d-clipplane";
 
 import "./styles/main.scss";
 
@@ -17,8 +17,8 @@ import InfoText from "./components/InfoText";
 import ModelSelector from "./components/ModelSelector";
 import Instructions from "./components/instructions/Instructions";
 import Footer from "./components/Footer";
-import VolumeViewerWrapper from "./components/volume-viewer/VolumeViewerWrapper.jsx";
-import { ControlsProvider } from "./context/controls-context.js";
+import VolumeViewerWrapper from "./VolumeViewer-package/components/VolumeViewerWrapper";
+import { ControlsProvider } from "./VolumeViewer-package/context/controls-context.js";
 
 export default function App() {
   const [controlsVisible, setControlsVisible] = useState(false);
