+++ src/VolumeViewer-package/Aframe/loader.js
