+++ src/VolumeViewer-package/Aframe/render-2d-clipplane.js
