+++ src/VolumeViewer-package/misc/VolumeSlice.js
