+++ src/VolumeViewer-package/shaders/ccvLibVolumeShader.js
