+++ src/assets/config.json
@@ -34,7 +34,7 @@
   "slices": 55,
   "x_spacing": 2,
   "y_spacing": 2,
-  "z_spacing": 1,
+  "z_spacing": 2,
 
   "volumePosition": "0 0 0",
   "volumeRotation": "-55 0 0",
